Bulk delete

+++ src/Umbraco.Cms.Api.Management/Controllers/User/DeleteUsersController.cs
@@ -1,5 +1,6 @@
 ﻿using Asp.Versioning;
 using Microsoft.AspNetCore.Mvc;
+using Umbraco.Cms.Core.Security;
 using Umbraco.Cms.Core.Services;
 using Umbraco.Cms.Core.Services.OperationStatus;
 
@@ -8,15 +9,20 @@ namespace Umbraco.Cms.Api.Management.Controllers.User;
 [ApiVersion("1.0")]
 public class DeleteUserController : UserControllerBase
 {
-    public DeleteUserController(IUserService userService) => _userService = userService;
+    public DeleteUserController(IUserService userService, IBackOfficeSecurityAccessor backOfficeSecurityAccessor)
+    {
+        _userService = userService;
+        _backOfficeSecurityAccessor = backOfficeSecurityAccessor;
+    }
 
     private readonly IUserService _userService;
+    private readonly IBackOfficeSecurityAccessor _backOfficeSecurityAccessor;
 
     [MapToApiVersion("1.0")]
     [HttpDelete("{id:guid}")]
     public async Task<IActionResult> DeleteUser(Guid id)
     {
-        UserOperationStatus result = await _userService.DeleteAsync(id);
+        UserOperationStatus result = await _userService.DeleteAsync(CurrentUserKey(_backOfficeSecurityAccessor), new HashSet<Guid>(){id});
 
         return result is UserOperationStatus.Success
             ? Ok()
